refactor(bff): adjust namings
