+++ apps/aurora-portal/src/client/routes/gardener/clusters/index.tsx
@@ -14,7 +14,7 @@ export const Route = createFileRoute("/gardener/clusters/")({
   component: RouteComponent,
   loader: async ({ context }) => {
     const clusters = await context.trpcClient?.gardener.getClusters.query()
-    const permissions = await context.trpcClient?.gardener.getAccessReviewInformation.query()
+    const permissions = await context.trpcClient?.gardener.getPermissions.query()
 
     return {
       clusters,
