+++ apps/aurora-portal/src/server/Gardener/routers/index.ts
@@ -1,12 +1,12 @@
 import { shootRouter } from "./shootRouter"
 import { cloudProfilesRouter } from "./cloudProfilesRouter"
-import { accessReviewRouter } from "./accessReviewRouter"
+import { permissionsRouter } from "./permissionsRouter"
 import { auroraRouter } from "../../trpc"
 
 export const gardenerRouters = {
   gardener: auroraRouter({
     ...shootRouter,
     ...cloudProfilesRouter,
-    ...accessReviewRouter,
+    ...permissionsRouter,
   }),
 }
